requirements.txt
# Core ML and Computer Vision
ultralytics==8.0.196
opencv-python==4.8.1.78
torch==2.0.1+cpu
torchvision==0.15.2+cpu

# GUI and Image Processing
Pillow==10.0.0
tkinter-tooltip==2.1.0

# Data handling
numpy==1.24.3
matplotlib==3.7.2

# Optional: For dataset download
kaggle==1.5.16
roboflow==1.1.1